snapshot: client-server wrangling
