+++ mlxclip_cli_test.go
