--- mlxclip_py.go
@@ -1,10 +1,0 @@
-// go:build mlxclip_python
-
-package embeddings
-
-import (
-	"embed"
-)
-
-//go:embed mlxclip_py.txt
-var mlxclip_fs embed.FS
