--- mlxclip_py.txt
@@ -1,22 +1,0 @@
-from mlx_clip import mlx_clip
-
-import sys
-import json
-
-if __name__ == "__main__":
-
-    model_dir = "/usr/local/src/mlx-examples/clip/mlx_model"
-    clip = mlx_clip(model_dir)
-
-    target = sys.argv[1]
-    input = sys.argv[2]
-    output = sys.argv[3]
-
-    with open(output, "w") as wr :
-
-        if target == "image":
-            image_embedding = clip.image_encoder(input)
-            json.dump(image_embedding, wr)
-        else :
-            text_embedding = clip.text_encoder(text)
-            json.dump(text_embedding, wr)
