+++ openclip_client.go
@@ -1,28 +1,30 @@
 package embeddings
 
+// https://github.com/mlfoundations/open_clip
+
 import (
 	"context"
 	"encoding/base64"
-	"fmt"
 	"net/url"
 	"strings"
 	"time"
 )
 
-func init() {
-	ctx := context.Background()
-	RegisterEmbedder[float32](ctx, "siglip-client", NewSigLIPLocalClientEmbedder)
-	RegisterEmbedder[float32](ctx, "siglip-client32", NewSigLIPLocalClientEmbedder)
-	RegisterEmbedder[float64](ctx, "siglip-client64", NewSigLIPLocalClientEmbedder)
-}
-
-type SigLIPLocalClientEmbedder[T Float] struct {
+// OpenCLIPEmbedder implements the `Embedder` interface using an OpenCLIP API endpoint to derive embeddings.
+type OpenCLIPEmbedder[T Float] struct {
 	Embedder[T]
 	client    *LocalClient
 	precision string
 }
 
-func NewSigLIPLocalClientEmbedder[T Float](ctx context.Context, uri string) (Embedder[T], error) {
+func init() {
+	ctx := context.Background()
+	RegisterEmbedder[float64](ctx, "openclip-client", NewOpenCLIPEmbedder)
+	RegisterEmbedder[float32](ctx, "openclip-client32", NewOpenCLIPEmbedder)
+	RegisterEmbedder[float64](ctx, "openclip-client64", NewOpenCLIPEmbedder)
+}
+
+func NewOpenCLIPEmbedder[T Float](ctx context.Context, uri string) (Embedder[T], error) {
 
 	u, err := url.Parse(uri)
 
@@ -37,28 +39,21 @@ func NewSigLIPLocalClientEmbedder[T Float](ctx context.Context, uri string) (Emb
 	if q.Has("client-uri") {
 		client_uri = q.Get("client-uri")
 	}
-	
-	cl, err := NewLocalClient(ctx, client_uri)
+
+	local_cl, err := NewLocalClient(ctx, client_uri)
 
 	if err != nil {
 		return nil, err
 	}
 
-	precision := "float32"
-
-	if strings.HasSuffix(u.Scheme, "64") {
-		precision = fmt.Sprintf("%s#as-float%d", precision, 64)
-	}
-
-	e := &SigLIPLocalClientEmbedder[T]{
-		client:    cl,
-		precision: precision,
+	e := &OpenCLIPEmbedder[T]{
+		client: local_cl,
 	}
 
 	return e, nil
 }
 
-func (e *SigLIPLocalClientEmbedder[T]) TextEmbeddings(ctx context.Context, req *EmbeddingsRequest) (EmbeddingsResponse[T], error) {
+func (e *OpenCLIPEmbedder[T]) TextEmbeddings(ctx context.Context, req *EmbeddingsRequest) (EmbeddingsResponse[T], error) {
 
 	cl_req := &LocalClientEmbeddingRequest{
 		Content: string(req.Body),
@@ -74,7 +69,7 @@ func (e *SigLIPLocalClientEmbedder[T]) TextEmbeddings(ctx context.Context, req *
 	return rsp, nil
 }
 
-func (e *SigLIPLocalClientEmbedder[T]) ImageEmbeddings(ctx context.Context, req *EmbeddingsRequest) (EmbeddingsResponse[T], error) {
+func (e *OpenCLIPEmbedder[T]) ImageEmbeddings(ctx context.Context, req *EmbeddingsRequest) (EmbeddingsResponse[T], error) {
 
 	data_b64 := base64.StdEncoding.EncodeToString(req.Body)
 
@@ -102,7 +97,7 @@ func (e *SigLIPLocalClientEmbedder[T]) ImageEmbeddings(ctx context.Context, req 
 	return rsp, nil
 }
 
-func (e *SigLIPLocalClientEmbedder[T]) localClientResponseToEmbeddingsResponse(req *EmbeddingsRequest, cl_rsp *LocalClientEmbeddingResponse) EmbeddingsResponse[T] {
+func (e *OpenCLIPEmbedder[T]) localClientResponseToEmbeddingsResponse(req *EmbeddingsRequest, cl_rsp *LocalClientEmbeddingResponse) EmbeddingsResponse[T] {
 
 	now := time.Now()
 	ts := now.Unix()
@@ -111,7 +106,7 @@ func (e *SigLIPLocalClientEmbedder[T]) localClientResponseToEmbeddingsResponse(r
 		CommonId:        req.Id,
 		CommonPrecision: e.precision,
 		CommonCreated:   ts,
-		CommonModel:     cl_rsp.Model,
+		CommonModel:     "openclip",
 	}
 
 	e64 := cl_rsp.Embeddings
